22.02 mk

+++ Module_00/ex00/megaphone.cpp
@@ -11,7 +11,7 @@ int	main(int argc, char **argv)
 		for (int i = 1; i < argc; i++)
 		{
 			std::string str = argv[i];
-			for (int f = 0; f < str.size(); f++)
+			for (long unsigned int f = 0; f < str.size(); f++)
 				std::cout << (char)toupper(str[f]);
 		}
 		std::cout << "\n";
